+++ src/Symfony/Component/Debug/Exception/ClassNotFoundException.php
@@ -28,5 +28,6 @@ class ClassNotFoundException extends FatalErrorException
             $previous->getLine(),
             $previous->getPrevious()
         );
+        $this->setTrace($previous->getTrace());
     }
 }