+++ src/Symfony/Component/Debug/Exception/ContextErrorException.php
@@ -15,22 +15,9 @@ namespace Symfony\Component\Debug\Exception;
  * Error Exception with Variable Context.
  *
  * @author Christian Sciberras <uuf6429@gmail.com>
+ *
+ * @deprecated since version 2.5, to be removed in 3.0.
  */
-class ContextErrorException extends \ErrorException
+class ContextErrorException extends HandledErrorException
 {
-    private $context = array();
-
-    public function __construct($message, $code, $severity, $filename, $lineno, $context = array())
-    {
-        parent::__construct($message, $code, $severity, $filename, $lineno);
-        $this->context = $context;
-    }
-
-    /**
-     * @return array Array of variables that existed when the exception occurred
-     */
-    public function getContext()
-    {
-        return $this->context;
-    }
 }